Simplified frontend UeberhangMandate code
